Add remote container definition
This change adds a definition for a VSCode remote container. The remote container ensures a consistent development environment on any system (assuming VSCode as the editor) which makes it possible for developers to use whichever OS they prefer and allows their development machines to be clean of project dependencies.

This is a standard container for Node and TypeScript with minor differences from the one defined one VSCode's own sample Node project.