Initial web service
This change adds the initial web service code and sets up linting for the project. The web service has a very basic HTTP endpoint to verify that it's up, connects to Redis, and sets up a very basic web socket listener.

To facilitate local testing the web service can be run with docker-compose.

This project will eventually sit next to a worker with both the web and worker projects operating as yarn workspaces.