Remove express
This change removes express since ws supports running as a standalone server.

The original examples I was using always showed ws running using an express or Node HTTP server but a closer look at the documentation shows that it can run on its own. This simplifies the code and also solves a problem: the root route was causing websocket connections to fail. The root route wasn't doing anything and so the fastest solution is to drop it.