Add utilities for working with player answers
This change adds two utilities for reading and collecting statistics about player answers.

A player has to answer the given question in a limited amount of time and is free to change their answer as much as they want within that time. The chosen answer - if any - at the end of the time limit is used as their answer for that question.

Answers are collected from web frontends and are stored in hash map in Redis keyed by player ID whose values are the chosen index into the answers array of the players. This ensures each web frontend can store the answers of their players independently of each other and that the worker can read all answers easily at once. The key to the hash map is the game ID and the current question round.