Implement worker job processing
This change implements the worker's job processing function. This function fetches the question, players, and starts the game. All errors are caught so that failures in one game do not affect other running games.

Player information is passed into the worker. It is expected to be an array of objects that contain the player ID and player name as keys.