Notify players that the game will start
This change adds two new notifications from the game worker: one that the game starting and who is participating and a count down before the first question is shown.

These notifications are needed for several reasons:

* The game start notification notifies each web frontend about which players are participating in the game. Since games are started by a single web frontend and that web frontend doesn't have access to the state of the others this information has to come from the worker as it's the only place that knows this information and can communicate it to all interested parties.

There might be a concern that with a large number of players we will be sending a large amount of data to each web frontend. If we allow names to contain 40 characters and IDs to be 16 bytes then the information for 1000 players only takes up about 100kb (40 characters in JavaScript is 80 bytes). This is a small amount of information to deposit in the single Redis pub/sub channel that the web frontends read from.

Next, this event also allows the client applications to update their state to let users know that their game is about to start.

Finally, the count down gives those users time to get ready for their first question.