Add object to abstract Redis operations
This change adds a `RedisDataStore` object that abstracts away the Redis operations performed by the web frontend. These operations include managing the player waiting area and the player answers.

Player answers are simply stored in a know hash keyed on game and round so that the worker can perform an `HGETALL` operation to get all player answers.

The waiting area is a bit more complex because multiple web frontends will be reading and writing to it at the same time.

The waiting area is managed as a Redis list of serialized player information. A web frontend can add a player to the waiting area by pushing that player's information to the end of that list. If the list is long enough to start a game the web frontend will remove that many players from the list and start a new game with those players.

The interesting aspect is that while one web frontend is removing players another may be adding players and subsequently also deciding to start a new game with them. To manage this we watch a lock key that will only allow players to be removed from the list if the lock hasn't been set since the list was read. The lock's value is set in the same Redis transaction as removing the players and so only one web frontend will be able to successfully remove players.

A web frontend that detects that it hasn't been able to successfully remove players will know that another web frontend got to it first. It's possible that the player it just added will not be part of the new game but that's ok. That player instead waits for enough players for the next game to start without issue.

A final aspect of the waiting area is that we don't attempt to remove players that have been waiting but subsequently disconnected from the game. Those players' information will be sent as valid game players but they will be eliminated from the first round since they won't provide an answer. This won't affect other players' experiences because they have no way of knowing if a different player answered incorrectly, failed to answer, or isn't connected.