Implement the web frontend
This change implements the web frontend. This server is responsible for connecting the various pieces that control the business logic and behavior of the service and is responsible for watching and distributing the events coming from the worker.

It has a periodic process that pings each client to make sure they are still available. If a client isn't available its socket is terminated to free resources.