Implement data channel
This change adds a `DataChannel` class which abstracts the communication between the application and the server. It raises various events for the incoming game messages, raises errors for the various error states, and manages the web socket.

The events raised by this object will be used to control the state machine of the application.