Build React application
This change begins moving the client towards React. The application is split into several different pages with a simple state machine that transitions between the pages.

Since this is on GitHub Pages this isn't using a build system and so I am using React without JSX - a very strange thing to do. Under normal circumstances this would be built into a bundled JavaScript file from a TypeScript or JavaScript + JSX source with Babel.