Improve the layout for all components
This change improves the layout for all components while fixing some bugs along the way.

Text size is increased across the board to make better use of space. Elements are separated, colored, and organized to make better use of space. The number of active players shows after the first question completes (since that's the point at which the game knows about who is playing).